+++ static/js/nzbhydra.js
@@ -40,9 +40,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config"
-                        }
+                        }]
                     }
                 }
             }
@@ -61,9 +61,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config (Auth)"
-                        }
+                        }]
                     }
                 }
             }
@@ -82,9 +82,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config (Searching)"
-                        }
+                        }]
                     }
                 }
             }
@@ -103,9 +103,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config (Categories)"
-                        }
+                        }]
                     }
                 }
             }
@@ -124,9 +124,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config (Downloader)"
-                        }
+                        }]
                     }
                 }
             }
@@ -145,9 +145,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Config (Indexers)"
-                        }
+                        }]
                     }
                 }
             }
@@ -166,9 +166,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System"
-                        }
+                        }]
                     }
                 }
             }
@@ -187,9 +187,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (Log)"
-                        }
+                        }]
                     }
                 }
             }
@@ -205,9 +205,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                     }],
                     resolve: {
                         loginRequired: loginRequiredStats,
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Stats"
-                        }
+                        }]
                     }
                     
                 }
@@ -224,9 +224,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         stats: ['loginRequired', 'StatsService', function (loginRequired, StatsService) {
                             return StatsService.get();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Stats"
-                        }
+                        }]
                     }
                 }
             }
@@ -244,9 +244,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                                 return response.indexerStatuses;
                             });
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Stats (Indexers)"
-                        }
+                        }]
                     }
                 }
             }
@@ -262,9 +262,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         history: ["StatsService", function(StatsService) {
                             return StatsService.getSearchHistory();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Stats (Searches)"
-                        }
+                        }]
                     }
                 }
             }
@@ -280,9 +280,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         downloads: ["StatsService", function (StatsService) {
                             return StatsService.getDownloadHistory();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Stats (Downloads)"
-                        }
+                        }]
                     }
                 }
             }
@@ -301,9 +301,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         askAdmin: ['loginRequired', '$http', function (loginRequired, $http) {
                             return $http.get("internalapi/askadmin");
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System"
-                        }
+                        }]
                     }
                 }
             }
@@ -319,9 +319,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (Updates)"
-                        }
+                        }]
                     }
                 }
             }
@@ -337,9 +337,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (Log)"
-                        }
+                        }]
                     }
                 }
             }
@@ -355,9 +355,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (Backup)"
-                        }
+                        }]
                     }
                 }
             }
@@ -373,9 +373,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (About)"
-                        }
+                        }]
                     }
                 }
             }
@@ -391,9 +391,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "System (Bug report)"
-                        }
+                        }]
                     }
                 }
             }
@@ -409,9 +409,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         safeConfig: ['loginRequired', 'ConfigService', function (loginRequired, ConfigService) {
                             return ConfigService.getSafe();
                         }],
-                        $title: function () {
-                            return "Search"
-                        }
+                        $title: ["$stateParams", function ($stateParams) {
+                            return "Search";
+                        }]
                     }
                 }
             }
@@ -427,9 +427,20 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                     },
                     resolve: {
                         loginRequired: loginRequiredSearch,
-                        $title: function () {
-                            return "Search results"
-                        }
+                        $title: ["$stateParams", function ($stateParams) {
+                            console.log($stateParams);
+                            var title = "Search results";
+                            var details; 
+                            if ($stateParams.title) {
+                                details = $stateParams.title;
+                            } else if ($stateParams.query) {
+                                details = $stateParams.query;
+                            }
+                            if (details) {
+                                title += " (" + details + ")";
+                            }
+                            return title;
+                        }]
                     }
                 }
             }
@@ -444,9 +455,9 @@ angular.module('nzbhydraApp').config(["$stateProvider", "$urlRouterProvider", "$
                         loginRequired: function () {
                             return null;
                         },
-                        $title: function () {
+                        $title: ["$stateParams", function ($stateParams) {
                             return "Login"
-                        }
+                        }]
                     }
                 }
             }
@@ -577,7 +588,7 @@ nzbhydraapp.run(["$rootScope", function ($rootScope) {
     $rootScope.$on('$stateChangeSuccess',
         function (event, toState, toParams, fromState, fromParams) {
             try {
-                $rootScope.title = toState.views[Object.keys(toState.views)[0]].resolve.$title();
+                $rootScope.title = toState.views[Object.keys(toState.views)[0]].resolve.$title[1](toParams);
             } catch(e) {
                 
             }
@@ -3406,8 +3417,6 @@ function ConfigService($http, $q, $cacheFactory) {
     }
 
     function maySeeAdminArea() {
-        console.log("MAYSEEADMINAREA IS SET TO TRUE FOR DEBUGGING");
-        return true;
         function loadAll() {
             var maySeeAdminArea = cache.get("maySeeAdminArea");
             if (!angular.isUndefined(maySeeAdminArea)) {